% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{download_if_url}
\alias{download_if_url}
\title{Resolve file or URL}
\usage{
download_if_url(path)
}
\arguments{
\item{path}{File path or URL.}
}
\description{
Resolve file or URL
}
\details{
If \code{path} is a local file path, assert that it
exists, then just return it. If \code{path} is a URL, see if it
has been downloaded before by checking if the file
\code{basename(path)} exists. If it exists, just return
\code{basename(path)}. Otherwise, download the file at
\code{path} to \code{basename(path)} and return
\code{basename(path)}.
}
\keyword{internal}
