% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_if_targz}
\alias{extract_if_targz}
\title{Decompress a gzipped tarball}
\usage{
extract_if_targz(path)
}
\arguments{
\item{path}{File path to gzipped tarball.}
}
\description{
Decompress a gzipped tarball
}
\details{
The contents of the tarball are extracted into a directory which has
directory name equals to the basename of the \code{path}. For example,
\code{a.tar.gz} containing \code{b.txt} will generate directory \code{./a/}
and file \code{./a/b.txt}.
}
