% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_by_num_overlaps}
\alias{filter_by_num_overlaps}
\title{Filter a GRanges object based on its number of overlaps with another}
\usage{
filter_by_num_overlaps(gr_to_filter, gr_to_overlap, filter_fn)
}
\arguments{
\item{gr_to_filter}{a GRanges object to filter}

\item{gr_to_overlap}{a GRanges object to overlap \code{to_filter} with}

\item{filter_fn}{A unary function which takes a numeric vector of length
\code{length(to_filter)}, where the i'th element of that vector is the
number of overlaps of the i'th row of \code{to_filter} with
\code{gr_to_overlap}. This function should then return a logical vector
(usually of length also equals \code{length(to_filter)}) with which to
subset (index) \code{to_filter} with.}
}
\value{
GRanges subset of \code{gr_to_filter}.
}
\description{
Filter a GRanges object based on its number of overlaps with another
}
