% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_bedtools_version}
\alias{get_bedtools_version}
\title{Get the current version of bedtools}
\usage{
get_bedtools_version(command = "bedtools --version")
}
\arguments{
\item{command}{Command string which produces a bedtools version string.}
}
\value{
Length-three numeric vector of major, minor, and patch versions
respectively.
}
\description{
Get the current version of bedtools
}
\keyword{internal}
