% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf2loci.R
\name{get_tf2loci}
\alias{get_tf2loci}
\title{Get a mapping of transcription factors to genomic loci}
\usage{
get_tf2loci(
 
    unibind_bed_dir = "https://unibind.uio.no/static/data/20210421/bulk_Robust/Homo_sapiens/damo_hg38_TFBS_per_TF.tar.gz",
  save_to_cache = TRUE,
  overwrite_cache = FALSE,
  read_from_cache = TRUE
)
}
\arguments{
\item{unibind_bed_dir}{File path or URL to a UniBind "TFBSs per TF (in BED
format)" directory or gzipped tarball (ending in \code{tar.gz}).}

\item{save_to_cache}{Should the results be cached? (\code{overwrite_cache}
modifies this behaviour.)}

\item{overwrite_cache}{If there was a pre-existing cache from the same
\code{unibind_bed_dir} with \code{save_to_cache}, should this new function
call overwrite that cached result? Note that different
\code{unibind_bed_dir} files are cached differently (they are
version-aware).}

\item{read_from_cache}{If there was a pre-existing cache from the same
\code{unibind_bed_dir}, should this new function call just read from that,
rather than try to download and parse it again?}
}
\value{
A GRangesList. The names of the list are transcription factor
names, and each element of the list is a GRanges recording the loci that
that transcription factor maps to, according to the \code{unibind_bed_dir}.
}
\description{
Get a mapping of transcription factors to genomic loci
}
\details{
\code{unibind_bed_dir} will be used directly if it is a local
directory. Otherwise, if it is a local gzipped tarball (determined by
checking for the file extension \code{tar.gz}), it will be extracted to a
local directory. If it is a remote (part of a URL) gzipped tarball, it will
downloaded then extracted to a local directory.

All UniBind TFBSs per TF BED files are available at their website,
https://unibind.uio.no/downloads/.
}
