% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_results_by_promoter.R
\name{group_results_by_promoter}
\alias{group_results_by_promoter}
\title{Group kallisto results by promoter}
\usage{
group_results_by_promoter(
  samples_to_covariates,
 
    transcripts_gtf_file = "http://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz",
 
    bedtools_genome_index_file = "https://raw.githubusercontent.com/arq5x/bedtools2/master/genomes/human.hg38.genome"
)
}
\arguments{
\item{transcripts_gtf_file}{File path or URL to a transcript GTF file in the
style of ENSEMBL.}

\item{bedtools_genome_index_file}{File path or URL to a "genome file" as
specified by bedtools.}

\item{sample_to_variates}{A \code{data.frame} in the same style as sleuth_prep.}
}
\description{
Group kallisto results by promoter
}
