% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sr2gr}
\alias{sr2gr}
\title{Convert sleuth results data.frame to GRanges}
\usage{
sr2gr(sr, mean_obs_col_name = "mean_obs")
}
\arguments{
\item{sr}{\code{data.frame} from \code{sleuth_results},}

\item{mean_obs_col_name}{Column name to find the quantity representing
mean log observations. For sleuth \code{gene_mode=TRUE}, this is
\code{"mean_obs"}, for sleuth \code{pval_aggregate=TRUE}, this is
\code{"sum_mean_obs_count"}.}
}
\value{
GRanges with qval, mean_obs, and target_id (usually a unique
TSS-associated ID) as metadata.
}
\description{
Convert sleuth results data.frame to GRanges
}
\details{
Rows with \code{NA} qvals are dropped. Whatever values are in the
column with the name \code{mean_obs_col_name} will be saved in the mcol
column named "mean_obs".
}
