% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tx2tss2gr}
\alias{tx2tss2gr}
\title{Convert tx2tss tibble to GRanges}
\usage{
tx2tss2gr(tx2tss)
}
\arguments{
\item{tx2tss}{tibble from get_tx2tss}
}
\value{
GRanges object with metadata column target_id.
}
\description{
Convert tx2tss tibble to GRanges
}
