% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{warn_if_outdated_bedtools}
\alias{warn_if_outdated_bedtools}
\title{Warn if current bedtools version is older than recommended}
\usage{
warn_if_outdated_bedtools(major = 2, minor = 28, patch = 0)
}
\arguments{
\item{major}{Numeric 'x' part of x.y.z.}

\item{minor}{Numeric 'y' part of x.y.z.}

\item{patch}{Numeric 'z' part of x.y.z.}
}
\value{
TRUE if current version < major.minor.patch, FALSE otherwise.
}
\description{
Warn if current bedtools version is older than recommended
}
\details{
If TRUE, this function produces a warning using the warning function.
}
\keyword{internal}
