% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gr.R
\name{aggregate_gr}
\alias{aggregate_gr}
\title{Aggregate GRanges-represented sleuth results using TF to loci mappings}
\usage{
aggregate_gr(
  gr,
  tf2loci,
  weight_fn = function(obs) log(obs + 1)/sum(log(obs + 1))
)
}
\arguments{
\item{gr}{GRanges representation of sleuth results obtained using
\code{sr2gr}.}

\item{tf2loci}{List of GRanges representing transcription factor to loci
bindings, obtained using \code{get_tf2loci}.}

\item{weight_fn}{Unary function to produce weights for Lancaster p-value
aggregation. This function should take a length-N numeric vector of
mean_obs, where N is the number of loci for that transcription factor.}
}
\value{
A tibble of results.
}
\description{
Aggregate GRanges-represented sleuth results using TF to loci mappings
}
\references{
Yi, L., Pimentel, H., Bray, N.L. et al. Gene-level
differential analysis at transcript-level resolution. Genome Biol
19, 53 (2018). https://doi.org/10.1186/s13059-018-1419-z
}
