% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gr.R
\name{aggregate_gr}
\alias{aggregate_gr}
\title{Aggregate GRanges-represented sleuth results using TF to loci mappings}
\usage{
aggregate_gr(
  gr,
  tf2loci,
  weight_fn = function(mcols) (1/mcols$tfs_overlapped)/sum(1/mcols$tfs_overlapped),
  filter_mapping_fn = function(n_rows) n_rows > median(n_rows)
)
}
\arguments{
\item{gr}{GRanges representation of sleuth results obtained using
\code{sr2gr}.}

\item{tf2loci}{List of GRanges representing transcription factor to loci
bindings, obtained using \code{get_tf2loci}.}

\item{weight_fn}{Unary function to produce weights for Lancaster p-value
aggregation. This function will receive \code{mcols(gr)}, and should return
a numeric vector of the same length as \code{length(gr)}. See details.}

\item{filter_mapping_fn}{Unary function to filter out TSS to TF mappings
which seem poorly supported by \code{tf2loci}. This function will receive
a numeric vector which represents the number of rows in tf2loci which
support each TSS to TF mapping, for each TSS (that is, each numeric vector
contains TSS to TF mappings for one TSS only). Then, the function should
return a logical vector representing which TSS to TF mappings should be
used.}
}
\value{
A tibble of results.
}
\description{
Aggregate GRanges-represented sleuth results using TF to loci mappings
}
\details{
This function will add a column "tfs_overlapped" to \code{mcols(gr)}
which is the number of transcription factors which each TSS in \code{gr}
maps to, if that column does not yet exists. So, you can use that column
in \code{weight_fn}.
}
\references{
Yi, L., Pimentel, H., Bray, N.L. et al. Gene-level
differential analysis at transcript-level resolution. Genome Biol
19, 53 (2018). https://doi.org/10.1186/s13059-018-1419-z
}
