% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf2loci.R
\name{get_tf2loci}
\alias{get_tf2loci}
\title{Get a mapping of transcription factors to genomic loci}
\usage{
get_tf2loci(
  unibind_bed_dir = "https://unibind.uio.no/static/data/bulk/pwm_tfbs_per_tf.tar.gz"
)
}
\arguments{
\item{unibind_bed_dir}{File path or URL to a UniBind "TFBs per TF (in BED
format)" directory or gzipped tarball (ending in \code{tar.gz}).}
}
\description{
Get a mapping of transcription factors to genomic loci
}
\details{
\code{unibind_bed_dir} will be used directly if it is a local
directory. Otherwise, if it is a local gzipped tarball (determined by
checking for the file extension \code{tar.gz}), it will be extracted to a
local directory. If it is a remote (part of a URL) gzipped tarball, it will
downloaded then extracted to a local directory.

All UniBind TFBSs per TF BED files are available at their website,
https://unibind.uio.no/downloads/.
}
