% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sr2gr}
\alias{sr2gr}
\title{Convert sleuth results data.frame to GRanges}
\usage{
sr2gr(sr)
}
\arguments{
\item{sr}{\code{data.frame} from \code{sleuth_results},}
}
\value{
GRanges with qval, mean_obs, and target_id (usually a unique
TSS-associated ID) as metadata.
}
\description{
Convert sleuth results data.frame to GRanges
}
\details{
Rows with \code{NA} qvals are dropped.
}
